Apply improvements Renato added to plugin organization
