+++ src/bp-members/js/blocks/member.js
@@ -1,162 +1,25 @@
 /**
  * WordPress dependencies.
  */
-const { registerBlockType } = wp.blocks;
-const { createElement, Fragment } = wp.element;
-const { Placeholder, Disabled, PanelBody, SelectControl, ToggleControl, Toolbar, ToolbarButton } = wp.components;
-const { InspectorControls, BlockControls } = wp.blockEditor;
-const { withSelect } = wp.data;
-const { compose } = wp.compose;
-const { ServerSideRender } = wp.editor;
-const { __ } = wp.i18n;
+const {
+	blocks: {
+		registerBlockType,
+	},
+	i18n: {
+		__,
+	},
+} = wp;
 
 /**
- * BuddyPress dependencies.
+ * Internal dependencies.
  */
-const { AutoCompleter } = bp.blockComponents;
-
-const AVATAR_SIZES = [
-	{
-		label: __( 'None', 'buddypress' ),
-		value: 'none',
-	},
-	{
-		label: __( 'Thumb', 'buddypress' ),
-		value: 'thumb',
-	},
-	{
-		label: __( 'Full', 'buddypress' ),
-		value: 'full',
-	},
-];
-
-const getSlugValue = ( item ) => {
-	if ( item && item.mention_name ) {
-		return item.mention_name;
-	}
-
-	return null;
-}
-
-const editMember = ( { attributes, setAttributes, bpSettings } ) => {
-	const { isAvatarEnabled, isMentionEnabled, isCoverImageEnabled } = bpSettings;
-	const { avatarSize, displayMentionSlug, displayActionButton, displayCoverImage } = attributes;
-
-	if ( ! attributes.itemID ) {
-		return (
-			<Placeholder
-				icon="admin-users"
-				label={ __( 'BuddyPress Member', 'buddypress' ) }
-				instructions={ __( 'Start typing the name of the member you want to feature into this post.', 'buddypress' ) }
-			>
-				<AutoCompleter
-					component="members"
-					slugValue={ getSlugValue }
-					ariaLabel={ __( 'Member\'s username', 'buddypress' ) }
-					placeholder={ __( 'Enter Member\'s username here…', 'buddypress' ) }
-					onSelectItem={ setAttributes }
-					useAvatar={ isAvatarEnabled }
-				/>
-			</Placeholder>
-		);
-	}
-
-	return (
-		<Fragment>
-			<BlockControls>
-				<Toolbar>
-					<ToolbarButton
-						icon="edit"
-						title={ __( 'Select another member', 'buddypress' ) }
-						onClick={ () =>{
-							setAttributes( { itemID: 0 } );
-						} }
-					/>
-				</Toolbar>
-			</BlockControls>
-			<InspectorControls>
-				<PanelBody title={ __( 'Settings', 'buddypress' ) } initialOpen={ true }>
-					<ToggleControl
-						label={ __( 'Display Profile button', 'buddypress' ) }
-						checked={ !! displayActionButton }
-						onChange={ () => {
-							setAttributes( { displayActionButton: ! displayActionButton } );
-						} }
-						help={
-							displayActionButton
-								? __( 'Include a link to the user\'s profile page under their display name.', 'buddypress' )
-								: __( 'Toggle to display a link to the user\'s profile page under their display name.', 'buddypress' )
-						}
-					/>
-
-					{ isAvatarEnabled && (
-						<SelectControl
-							label={ __( 'Avatar size', 'buddypress' ) }
-							value={ avatarSize }
-							options={ AVATAR_SIZES }
-							help={ __( 'Select "None" to disable the avatar.', 'buddypress' ) }
-							onChange={ ( option ) => {
-								setAttributes( { avatarSize: option } );
-							} }
-						/>
-					) }
-
-					{ isCoverImageEnabled && (
-						<ToggleControl
-							label={ __( 'Display Cover Image', 'buddypress' ) }
-							checked={ !! displayCoverImage }
-							onChange={ () => {
-								setAttributes( { displayCoverImage: ! displayCoverImage } );
-							} }
-							help={
-								displayCoverImage
-									? __( 'Include the user\'s cover image over their display name.', 'buddypress' )
-									: __( 'Toggle to display the user\'s cover image over their display name.', 'buddypress' )
-							}
-						/>
-					) }
-
-					{ isMentionEnabled && (
-						<ToggleControl
-							label={ __( 'Display Mention slug', 'buddypress' ) }
-							checked={ !! displayMentionSlug }
-							onChange={ () => {
-								setAttributes( { displayMentionSlug: ! displayMentionSlug } );
-							} }
-							help={
-								displayMentionSlug
-									? __( 'Include the user\'s mention name under their display name.', 'buddypress' )
-									: __( 'Toggle to display the user\'s mention name under their display name.', 'buddypress' )
-							}
-						/>
-					) }
-				</PanelBody>
-			</InspectorControls>
-			<Disabled>
-				<ServerSideRender block="bp/member" attributes={ attributes } />
-			</Disabled>
-		</Fragment>
-	);
-};
-
-const editMemberBlock = compose( [
-	withSelect( ( select ) => {
-		const editorSettings = select( 'core/editor' ).getEditorSettings();
-		return {
-			bpSettings: editorSettings.bp.members || {},
-		};
-	} ),
-] )( editMember );
+import editMemberBlock from './member/edit';
 
 registerBlockType( 'bp/member', {
 	title: __( 'Member', 'buddypress' ),
-
 	description: __( 'BuddyPress Member.', 'buddypress' ),
-
 	icon: 'admin-users',
-
 	category: 'buddypress',
-
 	attributes: {
 		itemID: {
 			type: 'integer',
@@ -179,6 +42,5 @@ registerBlockType( 'bp/member', {
 			default: true,
 		},
 	},
-
 	edit: editMemberBlock,
 } );
